Moved enums to separate module.
